requirements.txt
ddgs==9.5.2
fastapi==0.115.0
groq==0.11.0
httpx==0.27.2
python-dotenv==1.0.1
uvicorn==0.30.6
